requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

openai==1.6.1
tiktoken==0.3.3          # <- prebuilt wheel, no Rust needed

chromadb==0.4.21
sentence-transformers==2.2.2
huggingface_hub<1.0

PyPDF2==3.0.1
python-docx==1.1.0

pydantic==2.6.0
numpy<2
